requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
psycopg2-binary==2.9.9
pandas>=2.1.4
numpy>=1.26.0
folium==0.15.1
plotly==5.18.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0 